add a folder for binary search app
